project pr14

+++ controllers/cards.js
@@ -1,22 +1,18 @@
 const Card = require('../models/card');
+const NotFoundError = require('../errors/notFoundError');
+const ValidateError = require('../errors/validateError');
+const ForbiddenError = require('../errors/forbiddenError');
 
-const {
-  serverError,
-  dataError,
-  unfound,
-  ok,
-} = require('../errors');
-
-module.exports.getCards = async (req, res) => {
+module.exports.getCards = async (req, res, next) => {
   try {
     const cards = await Card.find({});
-    res.status(ok).send(cards);
+    res.status(200).send(cards);
   } catch (err) {
-    res.status(serverError).send({ message: 'На сервере произошла ошибка' });
+    next(err);
   }
 };
 
-module.exports.createCard = async (req, res) => {
+module.exports.createCard = async (req, res, next) => {
   const { name, link } = req.body;
   const owner = req.user._id;
   try {
@@ -24,38 +20,33 @@ module.exports.createCard = async (req, res) => {
     return res.status(201).send(card);
   } catch (err) {
     if (err.name === 'ValidationError') {
-      return res.status(dataError).send({
-        message: 'Переданы некорректные данные при создании карточки',
-      });
+      next(new ValidateError('Переданы некорректные данные при создании карточки'));
     }
-    return res
-      .status(serverError)
-      .send({ message: 'На сервере произошла ошибка' });
+    return next(err);
   }
 };
 
-module.exports.deleteCard = async (req, res) => {
+module.exports.deleteCard = async (req, res, next) => {
   try {
-    const card = await Card.findByIdAndRemove(req.params.id);
+    const owner = req.user._id;
+    const card = await Card.findById(req.params.id);
     if (!card) {
-      return res
-        .status(unfound)
-        .send({ message: 'Карточка по указанному _id не найдена' });
+      throw new NotFoundError('Карточка по указанному _id не найдена');
     }
-    return res.status(ok).send(card);
+    if (card.owner.toString() !== owner) {
+      throw new ForbiddenError('Недостаточно прав для удаление карточки');
+    }
+    await Card.findByIdAndRemove(req.params.id);
+    return res.status(200).send(card);
   } catch (err) {
     if (err.name === 'CastError') {
-      return res
-        .status(dataError)
-        .send({ message: 'Передан некорректный _id карточки' });
+      next(new ValidateError('Передан некорректный id карточки'));
     }
-    return res
-      .status(serverError)
-      .send({ message: 'На сервере произошла ошибка' });
+    return next(err);
   }
 };
 
-module.exports.likeCard = async (req, res) => {
+module.exports.likeCard = async (req, res, next) => {
   try {
     const card = await Card.findByIdAndUpdate(
       req.params.id,
@@ -63,24 +54,18 @@ module.exports.likeCard = async (req, res) => {
       { new: true },
     );
     if (!card) {
-      return res
-        .status(unfound)
-        .send({ message: 'Карточка по указанному _id не найдена' });
+      throw new NotFoundError('Карточка по указанному _id не найдена');
     }
-    return res.status(ok).send(card);
+    return res.status(200).send(card);
   } catch (err) {
     if (err.name === 'CastError') {
-      return res
-        .status(dataError)
-        .send({ message: 'Передан некорректный _id карточки' });
+      next(new ValidateError('Передан некорректный id карточки'));
     }
-    return res
-      .status(serverError)
-      .send({ message: 'На сервере произошла ошибка' });
+    return next(err);
   }
 };
 
-module.exports.dislikeCard = async (req, res) => {
+module.exports.dislikeCard = async (req, res, next) => {
   try {
     const card = await Card.findByIdAndUpdate(
       req.params.id,
@@ -88,19 +73,13 @@ module.exports.dislikeCard = async (req, res) => {
       { new: true },
     );
     if (!card) {
-      return res
-        .status(unfound)
-        .send({ message: 'Карточка по указанному _id не найдена' });
+      throw new NotFoundError('Карточка по указанному _id не найдена');
     }
-    return res.status(ok).send(card);
+    return res.status(200).send(card);
   } catch (err) {
     if (err.name === 'CastError') {
-      return res
-        .status(dataError)
-        .send({ message: 'Передан некорректный _id карточки' });
+      next(new ValidateError('Передан некорректный id карточки'));
     }
-    return res
-      .status(serverError)
-      .send({ message: 'На сервере произошла ошибка' });
+    return next(err);
   }
 };
